Did some massive work on the autosaving per-column and excellish table
git-svn-id: svn+ssh://scheduler.csc.calpoly.edu/repo@1802 13b268f7-8392-4c9d-bc7d-1e80a86bc2f9

+++ implementation/GWTView/src/edu/calpoly/csc/scheduler/view/web/client/table/columns/DeleteCell.java
@@ -3,45 +3,25 @@ package edu.calpoly.csc.scheduler.view.web.client.table.columns;
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.ClickHandler;
 import com.google.gwt.user.client.ui.Button;
-import com.google.gwt.user.client.ui.SimplePanel;
-import com.google.gwt.user.client.ui.Widget;
 
 import edu.calpoly.csc.scheduler.view.web.client.table.OsmTable;
 import edu.calpoly.csc.scheduler.view.web.client.table.OsmTable.IRowForCell;
 
-class DeleteCell extends SimplePanel implements OsmTable.Cell, OsmTable.EditingModeAwareCell {
-	boolean editing;
-	IRowForCell row;
+class DeleteCell extends OsmTable.Cell {
+	public interface DeleteStrategy {
+		void afterDelete();
+		boolean confirmHandler();
+	}
 	
-	Button deleteButton;
-	
-	DeleteCell(final IRowForCell row) {
-		this.row = row;
-		
-		editing = false;
-		
-		deleteButton = new Button("Delete", new ClickHandler() {
+	DeleteCell(final DeleteStrategy handler, final IRowForCell row) {
+		add(new Button("Delete", new ClickHandler() {
 			public void onClick(ClickEvent event) {
-				row.delete();
+				if (handler.confirmHandler()) {
+					row.delete();
+					handler.afterDelete();
+				}
+				event.stopPropagation();
 			}
-		});
-		
-		add(deleteButton);
+		}));
 	}
-	
-	@Override
-	public void enterEditingMode() {
-		assert(!editing);
-		editing = true;
-		deleteButton.setEnabled(false);
-	}
-
-	@Override
-	public void exitEditingMode() {
-		editing = false;
-		deleteButton.setEnabled(true);
-	}
-	
-	@Override
-	public Widget getCellWidget() { return this; }
 }
